tests: address PR review feedback on matcher fixtures
Flatten rule format (remove meta: wrapper), split monolithic fixture
files into themed files (scopes, logic, strings, parsing, span), make
each test small and focused with its own features, merge match_fixtures.py
into test_match_fixtures.py, and consolidate docs into the README.
