+++ tests/fixtures/matcher/dynamic/span.yml
@@ -1,0 +1,47 @@
+- name: span-window-contains-all
+  description: span-of-calls matches when all features fall within the window
+  options:
+    span size: 2
+  rules:
+    - name: span-resolve-add-veh
+      description: should match the span ending at the call that resolves AddVectoredExceptionHandler
+      scopes:
+        dynamic: span of calls
+      features:
+        - and:
+            - api: LdrGetDllHandle
+            - api: LdrGetProcedureAddress
+            - string: AddVectoredExceptionHandler
+  features: |
+    proc: sample.exe (ppid=2456, pid=3052)
+     thread: 3064
+      call: 10: api(LdrGetDllHandle)
+      call: 11: api(LdrGetProcedureAddress)
+      call: 11: string(AddVectoredExceptionHandler)
+      call: 12: api(RtlAddVectoredExceptionHandler)
+  expect:
+    matches:
+      span-resolve-add-veh:
+        - 11
+
+- name: span-window-too-small
+  description: span-of-calls does not match when the window is too small to contain all features
+  options:
+    span size: 2
+  rules:
+    - name: span-window-too-small
+      description: should not match because the configured span window does not include both APIs together
+      scopes:
+        dynamic: span of calls
+      features:
+        - and:
+            - api: LdrGetDllHandle
+            - api: RtlAddVectoredExceptionHandler
+  features: |
+    proc: sample.exe (ppid=2456, pid=3052)
+     thread: 3064
+      call: 10: api(LdrGetDllHandle)
+      call: 11: api(LdrGetProcedureAddress)
+      call: 12: api(RtlAddVectoredExceptionHandler)
+  expect:
+    matches: {}
